refactor index.rs
